refactor(host): module rearrangement, top level doc writing
