+++ crates/host/src/lib.rs
@@ -1,36 +1,67 @@
-//! wasmCloud host library
+//! # wasmCloud Host Runtime
+//!
+//! wasmCloud is an open source Cloud Native Computing Foundation (CNCF) project that
+//! enables teams to build, manage, and scale polyglot Wasm apps across any cloud, Kubernetes (K8s), or edge environments.
+//!
+//! This library provides the host runtime for wasmCloud, which is a platform for running WebAssembly (Wasm) components and plugins.
+//! It includes pluggable support for various extensions and integrations, allowing developers to customize and extend the host's functionality.
+//!
+//! ## Key Modules
+//!
+//! - **[`wasmbus`]**: This module contains the core implementation of the wasmCloud host functionality. It provides the [`Host`] struct and related configurations, which are central to running and managing Wasm components.
+//! - **[`oci`]**: Offers configuration and utilities for fetching OCI (Open Container Initiative) artifacts.
+//! - **[`policy`]**: Defines the [`PolicyManager`] trait for applying additional security policies on top of the wasmCloud host.
+//! - **[`registry`]**: Provides the [`crate::registry::RegistryCredentialExt`] extension trait for working with registry credentials.
+//! - **[`secrets`]**: Contains the [`crate::secrets::SecretsManager`] trait for securely fetching secrets from a secret store.
+//! - **[`store`]**: Defines the [`crate::store::StoreManager`] trait for managing configuration and data from a backing store.
+//! - **[`workload_identity`]**: Experimental module for workload identity implementations.
+//!
+//! ## Extending the Host
+//!
+//! The top-level modules in this crate expose implementable extension traits that allow developers to extend the host's functionality. These traits can be supplied to an embedded host using the [`crate::wasmbus::HostBuilder`].
+//!
+//! For example, you can implement custom policies, secrets management, or registry configurations to tailor the host to your specific needs.
+//!
+//! ## Getting Started
+//!
+//! To get started with wasmCloud, refer to the [`wasmbus`] module for the core host functionality. From there, you can explore the other modules to add extensions and integrations as needed.
+//!
+//! For more information, visit the [wasmCloud homepage](https://wasmcloud.com).
 
 #![warn(missing_docs)]
 #![forbid(clippy::unwrap_used)]
 
-/// wasmbus host
+/// [crate::wasmbus::Host] implementation and [crate::wasmbus::nats] implementations for the
+/// wasmCloud host extension traits
 pub mod wasmbus;
 
-/// OCI artifact fetching
+/// Configuration for OCI artifact fetching [crate::oci::Config]
 pub mod oci;
 
-/// wasmCloud policy service
+/// [crate::policy::PolicyManager] trait for layering additional security policies on top of the
+/// wasmCloud host
 pub mod policy;
 
-/// Common registry types
+/// [crate::registry::RegistryCredentialExt] extension trait for converting registry credentials
+/// into [wasmcloud_core::RegistryConfig]
 pub mod registry;
 
-/// Secret management
+/// [crate::secrets::SecretsManager] trait for fetching secrets from a secret store
 pub mod secrets;
 
-/// wasmCloud host metrics
+/// [crate::store::StoreManager] trait for fetching configuration and data from a backing store
+pub mod store;
+
+/// Implementation of OpenTelemetry metrics for wasmCloud, primarily using [wasmcloud_tracing]
 pub(crate) mod metrics;
 
-/// experimental workload identity
+/// experimental workload identity implementation
 pub mod workload_identity;
 
 pub use metrics::HostMetrics;
 pub use oci::Config as OciConfig;
 pub use policy::{HostInfo as PolicyHostInfo, PolicyManager, Response as PolicyResponse};
 pub use wasmbus::{Host as WasmbusHost, HostConfig as WasmbusHostConfig};
-pub use wasmcloud_core::{OciFetcher, RegistryAuth, RegistryConfig, RegistryType};
-
-pub use url;
 
 use std::collections::HashMap;
 use std::path::PathBuf;
@@ -40,6 +71,7 @@ use tokio::fs;
 use tracing::{debug, instrument, warn};
 use url::Url;
 use wascap::jwt;
+use wasmcloud_core::{OciFetcher, RegistryAuth, RegistryConfig, RegistryType};
 
 #[derive(PartialEq)]
 enum ResourceRef<'a> {
