+++ crates/host/src/store.rs
